Moved stdcxx.h to make clear it's for code using thrift.
